@@ -2,9 +2,9 @@ class SessionsController < ApplicationController
   
   include AppInstall
 
-  skip_before_filter :connect_to_db
-  skip_before_filter :require_account, only: [:create, :create_from_heroku, :login_heroku_app, :destroy]
-  skip_before_filter :verify_authenticity_token, only: [:create, :create_from_heroku]
+  skip_before_action :connect_to_db
+  skip_before_action :require_account, only: [:create, :create_from_heroku, :login_heroku_app, :destroy]
+  skip_before_action :verify_authenticity_token, only: [:create, :create_from_heroku]
 
   def create_from_heroku
     session[:heroku_access_token] = request.env['omniauth.auth']['credentials']['token']
