@@ -1,7 +1,7 @@
 class DocsController < ApplicationController
 
-  skip_filter :require_account
-  skip_filter :connect_to_db, unless: :valid_db_url?
+  skip_before_action :require_account
+  skip_before_action :connect_to_db, unless: :valid_db_url?
 
   def index
     @full_title = 'Documentation | Adminium'
