@@ -10,7 +10,7 @@ Adminium::Application.configure do
   config.action_controller.perform_caching = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
   config.static_cache_control = "public, max-age=31536000"
   
   # Compress JavaScripts and CSS
